requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.25.2
pytesseract==0.3.10
requests==2.31.0
streamlit==1.28.1
openai>=1.0.0
pydantic==2.5.0
pandas==2.1.3
jsonschema==4.20.0
opencv-python-headless>=4.8.0
openpyxl==3.1.2
python-docx==1.1.0
